requirements.txt
dj-database-url==0.3.0
Django==1.8.3
django-postgrespool==0.3.0
djangorestframework==3.1.3
gunicorn==19.3.0
psycopg2==2.6.1
SQLAlchemy==1.0.8
wheel==0.24.0
whitenoise==2.0.2
